refactor: :truck: Initial attempt to create a better project structure
