+++ boosty_downloader/src/application/download_manager/utils/path_sanitizer.py
