+++ boosty_downloader/src/infrastructure/boosty_api/__init__.py
@@ -1,0 +1,5 @@
+from .core.client import BoostyAPIClient
+
+__all__ = [
+    'BoostyAPIClient',
+]
