+++ boosty_downloader/src/infrastructure/boosty_api/core/client.py
@@ -8,10 +8,14 @@ from typing import TYPE_CHECKING
 
 from pydantic import ValidationError
 
-from boosty_downloader.src.boosty_api.models.post.extra import Extra
-from boosty_downloader.src.boosty_api.models.post.post import Post
-from boosty_downloader.src.boosty_api.models.post.posts_request import PostsResponse
-from boosty_downloader.src.boosty_api.utils.filter_none_params import filter_none_params
+from boosty_downloader.src.infrastructure.boosty_api.models.post.extra import Extra
+from boosty_downloader.src.infrastructure.boosty_api.models.post.post import Post
+from boosty_downloader.src.infrastructure.boosty_api.models.post.posts_request import (
+    PostsResponse,
+)
+from boosty_downloader.src.infrastructure.boosty_api.utils.filter_none_params import (
+    filter_none_params,
+)
 
 if TYPE_CHECKING:
     from collections.abc import AsyncGenerator
