+++ boosty_downloader/src/infrastructure/boosty_api/core/endpoints.py
