+++ boosty_downloader/src/infrastructure/boosty_api/models/post/post_data_types/post_data_file.py
