+++ boosty_downloader/src/infrastructure/boosty_api/utils/auth_parsers.py
