+++ boosty_downloader/src/infrastructure/loggers/base.py
