+++ boosty_downloader/src/infrastructure/loggers/failed_downloads_logger.py
