+++ boosty_downloader/src/infrastructure/loggers/logger_instances.py
@@ -2,8 +2,8 @@
 
 from pathlib import Path
 
-from boosty_downloader.src.loggers.base import RichLogger
-from boosty_downloader.src.loggers.failed_downloads_logger import FailedDownloadsLogger
+from boosty_downloader.src.infrastructure.loggers.base import RichLogger
+from boosty_downloader.src.infrastructure.loggers.failed_downloads_logger import FailedDownloadsLogger
 
 downloader_logger = RichLogger('Boosty_Downloader')
