+++ boosty_downloader/src/infrastructure/post_caching/post_cache.py
@@ -8,10 +8,10 @@ from sqlalchemy import String, create_engine, text
 from sqlalchemy.exc import DatabaseError, OperationalError
 from sqlalchemy.orm import DeclarativeBase, Mapped, Session, mapped_column, sessionmaker
 
-from boosty_downloader.src.download_manager.download_manager_config import (
+from boosty_downloader.src.application.download_manager.download_manager_config import (
     DownloadContentTypeFilter,
 )
-from boosty_downloader.src.loggers.base import RichLogger
+from boosty_downloader.src.infrastructure.loggers.base import RichLogger
 
 
 class Base(DeclarativeBase):
