+++ boosty_downloader/src/infrastructure/yaml_configuration/sample_config.py
