+++ test/unit/download_manager/ok_video_ranking_test.py
@@ -1,11 +1,11 @@
-from boosty_downloader.src.boosty_api.models.post.post_data_types.post_data_ok_video import (
-    OkVideoType,
-    OkVideoUrl,
-)
-from boosty_downloader.src.download_manager.utils.ok_video_ranking import (
+from boosty_downloader.src.application.download_manager.utils.ok_video_ranking import (
     RankingDict,
     get_best_video,
     get_quality_ranking,
+)
+from boosty_downloader.src.infrastructure.boosty_api.models.post.post_data_types.post_data_ok_video import (
+    OkVideoType,
+    OkVideoUrl,
 )
